requirements.txt
appdirs==1.4.2
attrs==16.3.0
beautifulsoup4==4.5.3
betamax==0.8.0
bs4==0.0.1
packaging==16.8
pyparsing==2.1.10
PyYAML==3.12
requests==2.13.0
six==1.10.0
